requirements.txt
aiofiles==24.1.0
aistudio-sdk==0.3.8
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
bce-python-sdk==0.9.57
beautifulsoup4==4.14.3
brotli==1.2.0
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorlog==6.10.1
cssselect==1.3.0
cssutils==2.11.1
einops==0.8.1
et_xmlfile==2.0.0
fastapi==0.128.0
ffmpy==1.0.0
filelock==3.20.1
fsspec==2025.12.0
ftfy==6.3.1
future==1.0.0
git-lfs==1.6
gradio==6.2.0
gradio_client==2.0.2
groovy==0.1.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.2.3
idna==3.11
imagesize==1.4.1
Jinja2==3.1.6
joblib==1.5.3
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
minio==7.2.20
modelscope==1.33.0
more-itertools==10.8.0
networkx==3.6.1
numpy==2.2.6
opencv-contrib-python==4.10.0.84
opencv-python==4.12.0.88
openpyxl==3.1.5
opt-einsum==3.3.0
orjson==3.11.5
packaging==25.0
paddleocr==3.3.2
paddlepaddle==3.2.2
paddlex==3.3.12
pandas==2.3.3
pillow==12.0.0
premailer==3.10.0
prettytable==3.17.0
protobuf==6.33.2
psutil==7.2.1
psycopg==3.3.2
psycopg-binary==3.3.2
py-cpuinfo==9.0.0
pyclipper==1.4.0
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.10.1
PyMuPDF==1.26.7
pypdfium2==5.2.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-multipart==0.0.21
python-ldap==3.4.4
pytz==2025.2
PyYAML==6.0.2
regex==2025.11.3
redis==5.2.0
requests==2.32.5
rich==14.2.0
ruamel.yaml==0.18.17
ruamel.yaml.clib==0.2.15
safehttpx==0.1.7
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
semantic-version==2.10.0
sentencepiece==0.2.1
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
soupsieve==2.8.1
starlette==0.50.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
tomlkit==0.13.3
tqdm==4.67.1
typer==0.21.0
typer-slim==0.20.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
urllib3==2.6.2
uvicorn==0.40.0
wcwidth==0.2.14
